requirements.txt
BeautifulSoup==3.2.1
Django==1.6.1
Fabric==1.8.0
MySQL-python==1.2.5
Pillow==2.4.0
South==0.8.3
Unipath==1.0
coverage==3.7.1
decorator==3.4.0
django-bootstrap-breadcrumbs==0.6.2
django-bootstrap-toolkit==2.15.0
django-braces==1.2.2
django-coverage==1.2.4
django-datatables-view==1.8
django-debug-toolbar==1.2
django-extensions==1.2.5
django-localflavor-br==1.0
django-passwords==0.2.0
django-test-extensions==0.14
django-widget-tweaks==1.3
ecdsa==0.11
ipython==1.1.0
mock==1.0.1
model-mommy==1.0
paramiko==1.14.0
pycrypto==2.6.1
selenium==2.41.0
six==1.6.1
splinter==0.6.0
sqlparse==0.1.11
wsgiref==0.1.2
